Final edits. See patch notes
